requirements.txt
qiskit
qiskit_ibm_runtime
matplotlib
pylatexenc
ipykernel
python-dotenv
qiskit_aer
#tweedledum  # requires 3.6 <= python <= 3.11  requires local build of tweedledum instead
sqlalchemy
alembic
